requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
anthropic
python-dotenv
PyMuPDF
